Adding sprites to be created by power-ups
Don't have the power-up blocks yet, but we might just go with the
question mark random power-up thing for those. Here we have Yogi Bear
driving a car, our flier with rocket wings, the bullets we shoot with
the Visitor power-up, and the sword for the guy to swing in dungeon
crawler thing.