Power-Ups 5
There are now 5 functioning power-ups, fully 1/2 of our stated goal.
Also, started debugging Win(), but I have yet to check it.
This is probably all I'm going to do tonight. Anything else I would do
is add the name of the power-up at the bottom of the screen, so this is
pretty much what we're going with for the demo in class tomorrow.